Add a pre-push hook for tag release
There is a `hook` folder with a file `pre-push`
which is a bash script (should run in linux or
git bash in Windows).
Copy this file to your local `.git/hooks` folder
and everytime you push a tag (e.g.
`git push origin v0.1.0`) it runs the `release.m`
script (so Matlab has to be installed an on the PATH)